+++ include/output/filter-out-mp4.h
@@ -1,5 +1,5 @@
  /* fre:ac - free audio converter
-  * Copyright (C) 2001-2018 Robert Kausch <robert.kausch@freac.org>
+  * Copyright (C) 2001-2019 Robert Kausch <robert.kausch@freac.org>
   *
   * This program is free software; you can redistribute it and/or
   * modify it under the terms of the "GNU General Public License".
@@ -30,8 +30,8 @@ namespace freac
 
 			Int			 frameSize;
 
-			Int			 totalSamples;
-			Int			 encodedSamples;
+			Int64			 totalSamples;
+			Int64			 encodedSamples;
 			Int			 delaySamples;
 
 			Buffer<unsigned char>	 outBuffer;
